+++ src/features/manageProfileSettings/hooks/useDragHandlers.test.ts
@@ -1,5 +1,9 @@
+import { setInitialGlobalState } from '@/test/__mocks__/store';
+
 import { DragEndEvent, DragOverEvent, DragStartEvent } from '@dnd-kit/core';
 import { renderHook } from '@testing-library/react';
+
+import { useManageProfileSettingsStore } from '@/store';
 
 import { ComponentType } from '../components/ProfileSettingsEditor/BaseComponent';
 import { useDragHandlers } from './useDragHandlers';
@@ -65,18 +69,26 @@ describe('useDragHandlers', () => {
     return dragEvent;
   };
 
+  beforeEach(() => {
+    setInitialGlobalState([
+      {
+        store: useManageProfileSettingsStore,
+        state: {
+          setUnusedComponents: mockSetUnused,
+          setSelectedComponents: mockSetSelected,
+        },
+      },
+    ]);
+  });
+
   it('sets state for drag start', () => {
     const { result } = renderHook(() =>
       useDragHandlers({
-        unused: [],
-        selected: [],
         startingList: ComponentType.unused,
         cancelDrag: mockCancelDrag,
         startDrag: mockStartDrag,
         endDrag: mockEndDrag,
         setStartingList: mockSetStartingList,
-        setSelected: mockSetSelected,
-        setUnused: mockSetUnused,
       }),
     );
 
@@ -90,15 +102,11 @@ describe('useDragHandlers', () => {
   it('set state for drag cancel', () => {
     const { result } = renderHook(() =>
       useDragHandlers({
-        unused: [],
-        selected: [],
         startingList: ComponentType.unused,
         cancelDrag: mockCancelDrag,
         startDrag: mockStartDrag,
         endDrag: mockEndDrag,
         setStartingList: mockSetStartingList,
-        setSelected: mockSetSelected,
-        setUnused: mockSetUnused,
       }),
     );
 
@@ -110,15 +118,11 @@ describe('useDragHandlers', () => {
   it('switches from unused to selected on dragging unused over selected', () => {
     const { result } = renderHook(() =>
       useDragHandlers({
-        unused: [],
-        selected: [],
         startingList: ComponentType.unused,
         cancelDrag: mockCancelDrag,
         startDrag: mockStartDrag,
         endDrag: mockEndDrag,
         setStartingList: mockSetStartingList,
-        setSelected: mockSetSelected,
-        setUnused: mockSetUnused,
       }),
     );
 
@@ -132,15 +136,11 @@ describe('useDragHandlers', () => {
   it('switches from selected to unused on dragging selected over unused', () => {
     const { result } = renderHook(() =>
       useDragHandlers({
-        unused: [],
-        selected: [],
         startingList: ComponentType.selected,
         cancelDrag: mockCancelDrag,
         startDrag: mockStartDrag,
         endDrag: mockEndDrag,
         setStartingList: mockSetStartingList,
-        setSelected: mockSetSelected,
-        setUnused: mockSetUnused,
       }),
     );
 
@@ -154,15 +154,11 @@ describe('useDragHandlers', () => {
   it('does not switch lists dragging selected over selected', () => {
     const { result } = renderHook(() =>
       useDragHandlers({
-        unused: [],
-        selected: [],
         startingList: ComponentType.selected,
         cancelDrag: mockCancelDrag,
         startDrag: mockStartDrag,
         endDrag: mockEndDrag,
         setStartingList: mockSetStartingList,
-        setSelected: mockSetSelected,
-        setUnused: mockSetUnused,
       }),
     );
 
@@ -177,15 +173,11 @@ describe('useDragHandlers', () => {
   it('does not switch lists dragging unused over unused', () => {
     const { result } = renderHook(() =>
       useDragHandlers({
-        unused: [],
-        selected: [],
         startingList: ComponentType.unused,
         cancelDrag: mockCancelDrag,
         startDrag: mockStartDrag,
         endDrag: mockEndDrag,
         setStartingList: mockSetStartingList,
-        setSelected: mockSetSelected,
-        setUnused: mockSetUnused,
       }),
     );
 
@@ -200,15 +192,11 @@ describe('useDragHandlers', () => {
   it('does not do anything when dragging over undroppable regions', () => {
     const { result } = renderHook(() =>
       useDragHandlers({
-        unused: [],
-        selected: [],
         startingList: ComponentType.unused,
         cancelDrag: mockCancelDrag,
         startDrag: mockStartDrag,
         endDrag: mockEndDrag,
         setStartingList: mockSetStartingList,
-        setSelected: mockSetSelected,
-        setUnused: mockSetUnused,
       }),
     );
 
@@ -223,15 +211,11 @@ describe('useDragHandlers', () => {
   it('reorders selected on drag end for selected and sets state', () => {
     const { result } = renderHook(() =>
       useDragHandlers({
-        unused: [],
-        selected: [],
         startingList: ComponentType.selected,
         cancelDrag: mockCancelDrag,
         startDrag: mockStartDrag,
         endDrag: mockEndDrag,
         setStartingList: mockSetStartingList,
-        setSelected: mockSetSelected,
-        setUnused: mockSetUnused,
       }),
     );
 
@@ -245,15 +229,11 @@ describe('useDragHandlers', () => {
   it('reorders unused on drag end for unused and sets state', () => {
     const { result } = renderHook(() =>
       useDragHandlers({
-        unused: [],
-        selected: [],
         startingList: ComponentType.unused,
         cancelDrag: mockCancelDrag,
         startDrag: mockStartDrag,
         endDrag: mockEndDrag,
         setStartingList: mockSetStartingList,
-        setSelected: mockSetSelected,
-        setUnused: mockSetUnused,
       }),
     );
 
@@ -267,15 +247,11 @@ describe('useDragHandlers', () => {
   it('does not reorder on drag end for undroppable regions and sets state', () => {
     const { result } = renderHook(() =>
       useDragHandlers({
-        unused: [],
-        selected: [],
         startingList: ComponentType.unused,
         cancelDrag: mockCancelDrag,
         startDrag: mockStartDrag,
         endDrag: mockEndDrag,
         setStartingList: mockSetStartingList,
-        setSelected: mockSetSelected,
-        setUnused: mockSetUnused,
       }),
     );